-Implemented double buffered (package buffer and transmission buffer) interfaces for both USB and BT. -USB and BT should now behave the same. Lossless communication is now available for USB. -Proper handling of switching between lossy and lossless communication. -Increased buffer sizes and types to handle larger commands. -Integration of acknowledgement reception and transmission in the normal transmission pipeline. -Various cleanup, small bug fixes and parameter changes. -Removed NACK transmission completely.

+++ src/tables/control_tables.h
@@ -35,7 +35,7 @@
 typedef const struct {
 	uint8_t header;
 	void (*cmd_response)(uint8_t**);
-	uint8_t nrb_payload;
+	uint16_t nrb_payload;
 	uint8_t nr_fields;
 	uint8_t field_widths[];
 } command_info;
